Move the ansi implementation to a separate module.
This cuts down deps on the main module.
