+++ windows/legacyconsole/doc.go
@@ -1,0 +1,8 @@
+/*
+package legacyconsole can be used to interact with the legacy terminal mode in Windows (pre-Win10).
+
+This package is included for historical reasons and is separate from the rest of the repo in order
+keep depdencies separate since go-ansiterm is maintained on a best-effort basis.
+*/
+
+package legacyconsole
