+++ windows/legacyconsole/go.mod
@@ -1,0 +1,5 @@
+module github.com/moby/term/windows/legacyconsole
+
+go 1.16
+
+require github.com/Azure/go-ansiterm v0.0.0-20210617225240-d185dfc1b5a1
